requirements.txt
aiogram>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.1.0
jinja2>=3.1.0
